change naming to better organize light source files
